+++ src-js/src/api-models/base.ts
@@ -264,7 +264,12 @@ export const enum ApiBlockType {
   Word = "WORD",
 }
 
-const LAYOUT_BLOCK_TYPES = new Set([
+/**
+ * Canonical Set of API block types counted as Layout* elements
+ *
+ * See: https://docs.aws.amazon.com/textract/latest/dg/layoutresponse.html
+ */
+export const LAYOUT_BLOCK_TYPES = new Set([
   ApiBlockType.LayoutFigure,
   ApiBlockType.LayoutFooter,
   ApiBlockType.LayoutHeader,